pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "physicsnemo_curator"
authors = [
    {name = "NVIDIA Corporation"}
]
description = "Data curation tools for PhysicsNeMo"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
dependencies = [
    "numpy>=1.26.4",
    "pyvista>=0.44.2",
    "vtk>=9.3.1",
    "hydra-core>=1.3",
    "zarr~=2.18",
    "numcodecs>=0.13.1",
    "tqdm>=4.67.1",
    "lasso-python==2.0.3",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "black",
    "ruff",
    "interrogate",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-xdist",
    "debugpy<1.9.0",
]

[tool.setuptools.dynamic]
version = {attr = "physicsnemo_curator.__version__"}

[tool.setuptools.packages.find]
where = ["."]
include = ["physicsnemo_curator", "physicsnemo_curator.*"]

[tool.ruff]
lint.select = ["E", "F", "S", "I", "PERF"]
lint.fixable = ["I"]
lint.ignore = ["E501", "S311", "S603", "S607"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"test/*.py" = ["S101"]
"tests/ci_tests/header_check.py" = ["S603", "S607"]

[tool.pytest.ini_options]
addopts = "-v --cov=physicsnemo_curator --cov-report=term-missing"
testpaths = ["tests"]
python_files = "test_*.py"
